Rename skunk-sharp-json → skunk-sharp-circe; use skunk-circe 1.0.0
Renamed the module to match its actual dependency shape. The previous
naming implied "JSON support" in the abstract; this module is
specifically the skunk-circe integration (tags for Postgres json /
jsonb mapped to `io.circe.Json`). A future "no-circe" JSON variant
could ship separately.

Also: skunk-circe 1.0.0 stable DOES exist (I misread the earlier
metadata listing and skipped to milestones). Removed the inline
`Codec.simple` codec — skunk-circe's `json` / `jsonb` are what we
want.

Paths:
  modules/json              → modules/circe
  skunk.sharp.json.*        → skunk.sharp.circe.*
  artefact skunk-sharp-json → skunk-sharp-circe
  tests/.../tests/JsonbSuite keeps its name (Jsonb-focused content)

Tests: 10 unit + 4 integration — all green.
