+++ build.sbt
@@ -15,6 +15,7 @@ ThisBuild / tlJdkRelease   := Some(17)
 ThisBuild / tlFatalWarnings := true
 
 val skunkV           = "1.0.0"
+val skunkCirceV      = "1.0.0"
 val circeV           = "0.14.15"
 val catsEffectV      = "3.7.0"
 val munitV           = "1.2.4"
@@ -24,7 +25,7 @@ val testcontainersV  = "0.44.1"
 val dumboV           = "0.9.0-SNAPSHOT"
 val otel4sV          = "0.16.0"
 
-lazy val root = tlCrossRootProject.aggregate(core, iron, json, tests)
+lazy val root = tlCrossRootProject.aggregate(core, iron, circe, tests)
 
 lazy val core = project
   .in(file("modules/core"))
@@ -50,14 +51,14 @@ lazy val iron = project
     )
   )
 
-lazy val json = project
-  .in(file("modules/json"))
+lazy val circe = project
+  .in(file("modules/circe"))
   .dependsOn(core)
   .settings(
-    name := "skunk-sharp-json",
+    name := "skunk-sharp-circe",
     libraryDependencies ++= Seq(
+      "org.tpolecat"  %% "skunk-circe"       % skunkCirceV,
       "io.circe"      %% "circe-core"        % circeV,
-      "io.circe"      %% "circe-parser"      % circeV,
       "org.scalameta" %% "munit"             % munitV           % Test,
       "org.typelevel" %% "munit-cats-effect" % munitCatsEffectV % Test
     )
@@ -65,7 +66,7 @@ lazy val json = project
 
 lazy val tests = project
   .in(file("modules/tests"))
-  .dependsOn(core, iron, json)
+  .dependsOn(core, iron, circe)
   .enablePlugins(NoPublishPlugin)
   .settings(
     name := "skunk-sharp-tests",
