+++ modules/circe/src/main/scala/skunk/sharp/circe/PgJsonb.scala
@@ -1,4 +1,4 @@
-package skunk.sharp.json
+package skunk.sharp.circe
 
 import skunk.sharp.TypedExpr
