+++ modules/circe/src/test/scala/skunk/sharp/circe/JsonbSuite.scala
@@ -1,4 +1,4 @@
-package skunk.sharp.json
+package skunk.sharp.circe
 
 import io.circe.Json as CirceJson
 import skunk.sharp.dsl.*
